Fix WHO_AM_I read & refactor i2c calls
Refactored i2c read and write calls to their own functions. Definitely
still naively using the i2c-dev interface. There seems to be a more
ideal way to read and write by using an ioctl call. There are not many
examples online though and the kernel documentation is unclear on how
this should be used.