Fix read and write pointer bug
Fixed a bug where a unique pointer was type uint8_t instead of
uint8_t[]. Also fixed a bug where read and write were checked for the
wrong values. Discovered a bug where the read transaction doesn't work
because the delay between write and read is too long. Need to use the
proper i2c-dev ioctl to do this correctly.