Fix i2c calls to use ioctl mixed transaction (#7)
Previously, basic calls to write and read file descriptors were used.
Now the ioctl call is used which allows for a write and read to occur in
one transaction. This fixes the bug where you cannot read from a
selected register on a device.
Also added the datasheet and configuration manual for the MPL3115A2. This way
offline work can be done easier.