Implement some extreme basics for LSM9DS1
Implemented some basic functionality that the LSM9DS1 class is going to
have to provide. Really good chance it doesn't work at all, and even if
it does the interpretation of the data currently is worthless. And on
top of that the readings will be all over the place. The accelerometer
readings will be good in the long run but in the short run will be
jittery as can be. The gyrometer readings are the opposite. They are
accurate in short periods but they are going to drift off way too much.
And I have no clue how the magnetometer readings will be. Some form of
filter (complementary/kahlman) should be implemented to handle the noise
in these "signals".