Create a data-server using zeromq for ipc
Created a data-server executable that uses zeromq to send data to anyone
requesting it. Added an examples folder with a python client
implementation for interacting with this data-server. Finally, I pivoted
this repo's purpose again. I think our overall system design has finally
settled and this repo is going to just consist of data collection from
i2c devices on our beaglebone black. A python application will actually
control the UAV and communicate with the ground control station. The
python application would act as a client for the data-server created here
and would read data when needed by sending any request to the data-server.
The server will respond with temperature and altitude data once it gets
any request.